update mudlet-mapper.xml
I think this fixes an issue whereby mmp.startup() checks for if not mmp.firstRun when trying to initialise. I changed to if mmp.firstRun == false as a check, since an unitialised variable has the nil value and matches against "not" checks.

This seems to fix the issue with my script on initialisation in Mudlet 3.00 Epsilon that I know some others were having too.