Preparations for cron CI runs
This commit prepares for the ability to run CI jobs on a cron instead of
on every PR. That way we won't publish a new version with every merged
PR, but have one release on a regular schedule.

Te way it works is that it compares a checksum of the currently released
script with the script in the repository (leaving out the version line).
If those differ, a new release is created.