Lusternia fixes
Added a check to rift and pathway triggers so they don't horribly break if the game's inbuilt pathfinding is used and sends you through one. Might be possible to detect these with game-side changes, but for now this was the best solution to it clearing special exits and potentially creating erroneous ones.
Rifts and pathways created by the mapper script will now have weights of 20 and 15 respectively.
Updated one of the scent trigger lines.