Stop speedwalk for gmcp.Room.WrongDir in StickMUD
The purpose of this game specific addition is to stop a speedwalk when the game is StickMUD and a Room.WrongDir GMCP event is triggered.  The typical use case for this is when an NPC is blocking the path, which is a common way we slow down newbies from entering areas that are too tough for them.  One could test this by trying to go to Demon World on the mapper in StickMUD.  You should stop at the Guard and could not go "in".